+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Drop legacy call support (URL in `options`).
 - Rewrite tests to async/await.
 - Deps bump & cleanup.
+- Input stream will now be closed by default. If you want to keep it open, add `{ keepOpen: true }`.
 
 
 ## [5.0.0] - 2019-09-14